+++ fflas-ffpack/ffpack/ffpack_charpoly_kgfastgeneralized.inl
@@ -5,20 +5,20 @@
  *
  * Written by Clement Pernet <Clement.Pernet@imag.fr>
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library FFLAS-FFPACK.
- * 
+ *
  * FFLAS-FFPACK is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -462,7 +462,7 @@ namespace FFPACK {
 						FFLAS::fcopy (F, r, tmp+i*r, 1, C4 + i*lda, 1);
 					}
 					for (int i = int(N-1); i >= (int) (N -mu-r); --i)
-						FFLAS::fcopy (F, r, C+i*lda, 1, C+(i-mc+r)*lda, 1);
+						FFLAS::fcopy (F, r, C+i*(int)lda, 1, C+((size_t)i-mc+r)*lda, 1);
 #ifdef LB_DEBUG
 					std::cerr<<"..done"<<std::endl;
